Make FIFO files world readable on Android.
